Added loading functions for images (ImageLoader.cpp, ImageLoader.h). Now the engine supports '.bmp' and '.tga' files (24bit, 32bit). Added some bmp and tga files for test in the 'images' directory.
Added '.mtl' support for '.obj' files.
The current interpolation technique is linear.
Added some testfiles in the 'models' directory.

Changed the 'DrawTriangleFilled', 'ApplyLight' and 'DrawModel' functions in the 'Renderer.cpp' file.

Added helper structs in the 'misc' directory: light, material, texture

Now the Face stores the material name and the texture coordinates.

Now the Model stores the texture coordinates and the face material.